+++ src/components/homepage/NewArrival.jsx
@@ -1,6 +1,8 @@
 import React, { useEffect, useState } from 'react'
 import { Link } from 'react-router-dom'
 import axios from 'axios';
+import toast from 'react-hot-toast';
+import Cookies from 'js-cookie';
 
 const NewArrival = () => {
     const [products, setProducts] = useState([])
@@ -8,12 +10,45 @@ const NewArrival = () => {
     const [error, setError] = useState('')
     const [success, setSuccess] = useState(false)
 
+    // Add to cart if Logged in
+    const addCart = async (product) => {
+        const token = Cookies.get('token');
+        if (token) {
+            try {
+                setLoading(true);
+                const { data } = await axios.post(
+                    `${process.env.REACT_APP_BASE_URL}/api/v1/cart/addCart`,
+                    {
+                        productId: product.Id,
+                        qty: product.min_qty
+                    },
+                    {
+                        headers: {
+                            Authorization: `Bearer ${token}`,
+                            'Content-Type': 'application/json',
+                        },
+                    }
+                );
+                setLoading(false);
+            } catch (error) {
+                console.log(error);
+            }
+        }
+    };
+
+    const addCartPromise = (product) => {
+        toast.promise(addCart(product), {
+            loading: 'Adding to cart...',
+            success: 'Added to cart',
+            error: 'Could not add to cart',
+        });
+    }
+
     const fetchProducts = async () => {
         setLoading(true)
         try {
             const res = await axios.get(`${process.env.REACT_APP_BASE_URL}/api/v1/allProduct?categoryId=&productId=&maincateId=&search=gmail&limit&skip`)
             setProducts(res.data.proData)
-            // console.log(res.data.proData)
             setLoading(false)
             setSuccess(true)
         } catch (error) {
@@ -27,18 +62,33 @@ const NewArrival = () => {
         fetchProducts()
     }, [])
 
-
     const addToCart = (product) => {
         const cart = sessionStorage.getItem('cart') ? JSON.parse(sessionStorage.getItem('cart')) : [];
-        cart.push(product);
-        sessionStorage.setItem('cart', JSON.stringify(cart));
+        const existingProduct = cart.find((item) => item.Id === product.Id);
+        const Quantity = product.min_qty;
+        if (existingProduct) {
+            alert('Product already in cart');
+            for (let i = 0; i < cart.length; i++) {
+                if (cart[i].Id === product.Id) {
+                    const oldQty = cart[i].min_qty;
+                    cart[i].min_qty = Quantity + oldQty;
+                    sessionStorage.setItem('cart', JSON.stringify(cart));
+                }
+            }
+        }
+        else {
+            cart.push(product);
+            toast.success("Product added to cart",{
+                duration:800,
+            });
+            sessionStorage.setItem('cart', JSON.stringify(cart));
+
+        }
     }
-
-
     return (
         <section id="product1" className='section-p1'>
             <h2>New Arrivals</h2>
-            <p>Newly added or coming soon accounts</p>
+            <p>Most selling Accounts</p>
             <div className='pro-container'>
                 {
                     products.map((product) => (
@@ -47,18 +97,18 @@ const NewArrival = () => {
                             <div className='des'>
                                 <span>{product.type}</span>
                                 <h4>{product.banner_title}</h4>
-                                {/* <div className='star'> */}
-                                    {/* <i className='fa fa-star'></i> */}
-                                    {/* <i className='fa fa-star'></i> */}
-                                    {/* <i className='fa fa-star'></i> */}
-                                    {/* <i className='fa fa-star'></i> */}
-                                    {/* <i className='fa fa-star'></i> */}
+                                <div className='star'>
+                                    <i className='fa fa-star'></i>
+                                    <i className='fa fa-star'></i>
+                                    <i className='fa fa-star'></i>
+                                    <i className='fa fa-star'></i>
+                                    <i className='fa fa-star'></i>
                                     {/* <i className='fa fa-star-half-o'></i> */}
-                                {/* </div> */}
+                                </div>
                                 <h4><span>Price:</span> $ {product.price}</h4>
                             </div>
-                            <Link to='#' className='btn'><i class="fa-solid fa-cart-shopping cart" onClick={()=>{
-                                addToCart(product)
+                            <Link to='#' className='btn'><i class="fa-solid fa-cart-shopping cart" onClick={() => {
+                                Cookies.get('token') ? addCartPromise(product) : addToCart(product)
                             }}></i></Link>
                         </div>
                     ))
